# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCII Initial Scaffolding'"}],"paths":{"/api/countries":{"get":{"tags":["country-controller"],"operationId":"getAllCountries","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}},"post":{"tags":["country-controller"],"operationId":"postCountries","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DtoRequestNewContries"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/{name}/name":{"get":{"tags":["country-controller"],"operationId":"getCountryByName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CountryDTO"}}}}}}},"/api/countries/{language}/language":{"get":{"tags":["country-controller"],"operationId":"getCountrysByLanguage","parameters":[{"name":"language","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/{continent}/continent":{"get":{"tags":["country-controller"],"operationId":"getCountrysByContinent","parameters":[{"name":"continent","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"/api/countries/most-borders":{"get":{"tags":["country-controller"],"operationId":"getCountryWithMostBorders","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CountryDTO"}}}}}}}},"components":{"schemas":{"DtoRequestNewContries":{"type":"object","properties":{"amountOfCountryToSave":{"type":"integer","format":"int32"}}},"CountryDTO":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"}}}}}}